refactor: code completion
